Added: Dataset labeling
There are other csf files I later found in the website. These, as opposed to the ML ones, do contain the flow details for each label.
Based on that I label my own packets, its not perfect but should cover some 90+% of the cases.

Ammended: Some sanity checking for dates.