🎇Working: Sampled and labeled flows.
This commit introduces a fully functioning sampling with labels. These will be stored in `data/precalc_windows` and be used for training.
This means that our approach is to take a monolithic csv file (e.g. `data/Wednesday.csv`) that includes labels and using the `./sampleddetection/environment/models.py::Environment` class we create csv files that correspond to a selection of Window Skip (WS) and window length (WL).
Take into consideration that this workflow is specific to testing Decision Trees without RL aid, and that RL will necessarily be done on RAM.
Nonetheless, the upcoming RL flow will be implemented with the same fundamental classes.