Changed: Approach to sampling
Most important: Doubled down on the approach of first taking the
window_lenth nap right after the action so that it is the freshest
decision.
Also doubled down on the approach that the enviornment should keep state
of the window skip and length. Which reduces the responsibility of the
sampler and frankly makes it feel like it could be made into a simple
function. Maybe later though. Lets first see how things perform